refactor(github): move github identity helper
