+++ tests/integrations/test_github_login.py
@@ -4,7 +4,7 @@ import pytest
 
 from integrations.github import login as github_login
 from integrations.github.login import GitHubLoginResult
-from integrations.github_identity import saved_github_username
+from integrations.github.identity import saved_github_username
 from integrations.github_mcp import GitHubMCPValidationResult
 from integrations.github_mcp_oauth import GitHubDeviceToken
